Implement custom array size and speed multiplier

+++ Sorting-Visualiser-main/scripts/helper.js
@@ -1,7 +1,7 @@
 "use strict";
 class Helper {
     constructor(time, list = []) {
-        this.time = parseInt(400/time);
+        this.time = parseFloat(400/time);
         this.list = list;
     }
